User listings refactoring
