+++ frontend/src/app/main.tsx
@@ -10,15 +10,15 @@ import './styles/index.css'
 import PrivateRoute from '@/app/routes/private-routes.tsx'
 
 import { AccountPage } from '@/pages/account/account-page.tsx'
+import { ChatPage } from '@/pages/account/chat/chat-page.tsx'
+import { FavoritesPage } from '@/pages/account/favorites/favorites-page.tsx'
+import { ProfilePage } from '@/pages/account/profile/profile-page.tsx'
+import { SettingsPage } from '@/pages/account/settings/settings-page.tsx'
+import { UserListingsPage } from '@/pages/account/user-listings/user-listings-page.tsx'
 import { CatalogPage } from '@/pages/catalog/catalog-page.tsx'
-import { ChatPage } from '@/pages/chat/chat-page.tsx'
-import { FavoritesPage } from '@/pages/favorites/favorites-page.tsx'
 import { HomePage } from '@/pages/home/home-page'
 import { AccountLayout, RootLayout } from '@/pages/layouts'
 import { ProductPage } from '@/pages/product/product-page.tsx'
-import { ProfilePage } from '@/pages/profile/profile-page.tsx'
-import { SettingsPage } from '@/pages/settings/settings-page.tsx'
-import { UserListingsPage } from '@/pages/user-listings/user-listings-page.tsx'
 import { PRIVATE_PAGES, PUBLIC_PAGES } from '@/shared/constants'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
