+++ frontend/src/pages/account/account-page.tsx
@@ -1,5 +1,5 @@
 import { AccountPageMobile } from '@/pages/account/account-page-mobile.tsx'
-import { UserListingsPage } from '@/pages/user-listings/user-listings-page.tsx'
+import { UserListingsPage } from '@/pages/account/user-listings/user-listings-page.tsx'
 import { useMediaQuery } from '@/shared/library/hooks'
 
 export const AccountPage = () => {
