+++ frontend/src/pages/account/chat/chat-page.tsx
@@ -1,4 +1,4 @@
-import { ChatWrapper } from '@/features/chat/ui/chat-wrapper'
+import { ChatWrapper } from '@/features/chat/ui/chat-wrapper.tsx'
 
 export const ChatPage = () => {
   return (
