+++ frontend/src/pages/account/favorites/favorites-page.tsx
