+++ frontend/src/pages/account/welcome/welcome-page.tsx
